refactor: extraire la logique pure de App.tsx vers src/lib (#17)
Premier découpage de App.tsx (monolithique, 3222 lignes) en isolant la
logique pure et testable dans src/lib :

- types.ts    : VideoFile, Subtitle, Playlist, TMDB_GENRES
- utils.ts    : srt2vtt, isPersonalVideo, getCleanTitle, formatSize,
                formatDuration, getResolution
- grouping.ts : groupVideos (regroupement séries + sagas TMDB)
- sorting.ts  : isGroupWatched, sortAlphabetical, sortRecommendations,
                filterAndSortVideos

App.tsx ne fait plus qu'orchestrer l'état/UI et importe ces modules
(3222 -> 2920 lignes). Cette structure src/lib débloque #14 (tests) et
#12 (docs).

Corrige aussi un bug latent : le memo filteredAndSortedVideos lisait
watchedVideos sans le déclarer en dépendance (tri "vues à la fin" non
rafraîchi).
